+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 19/02/2020
+- Add Areas v2 router, which syncs areas of interest with subscriptions from the GFW Subscriptions API MS. For more information about this feature, check the docs.
+
 # 16/12/2019
 - Fix issue where creating areas would not save the thumbnail.